Implement RaiseError initialization arguments
This change implements a RaiseError argument when initializing
Mustache which will raise a `Mustache.TemplateMissing` if a
template (or partial) is missing from the file system, which is
that last resort in the lookup order.

The argument excepts a boolean for turning the feature on/off,
default is on.

When turned off, the error will be silenced and an empty string
will be returned for the missing template.